JOb-JVAI/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
pytz==2023.3
python-dateutil==2.8.2
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
cryptography==41.0.7
pyotp==2.9.0
qrcode==7.4.2
pytest==7.4.3
pytest-django==4.7.0
faker==20.1.0
